pyproject.toml
[project]
name = "mcp-webfetch"
version = "0.1.0"
description = "WebFetch MCP Server - Fetch web pages and answer questions about their content"
readme = "README.md"
license = {text = "MIT"}
requires-python = ">=3.13"
authors = [
    { name = "NimbleBrain Inc", email = "hello@nimblebrain.ai" },
]
keywords = ["mcp", "web", "fetch", "markdown", "nimblebrain"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.13",
    "Programming Language :: Python :: 3.14",
]
dependencies = [
    "aiohttp>=3.11.0",
    "anthropic>=0.49.0",
    "fastmcp>=2.14.0",
    "markdownify>=0.14.1",
    "python-dotenv>=1.0.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/mcp_webfetch"]

[tool.ruff]
target-version = "py313"
line-length = 100

[tool.ruff.lint]
select = [
    "E",   # pycodestyle errors
    "W",   # pycodestyle warnings
    "F",   # pyflakes
    "I",   # isort
    "B",   # flake8-bugbear
    "C4",  # flake8-comprehensions
    "UP",  # pyupgrade
]
ignore = [
    "E501",  # line too long - handled by formatter
    "B008",  # do not perform function calls in argument defaults
]

[tool.ruff.format]
quote-style = "double"
indent-style = "space"

[tool.mypy]
python_version = "3.13"
strict = true
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
ignore_missing_imports = true
disable_error_code = ["unused-coroutine", "misc"]

[dependency-groups]
dev = [
    "mypy>=1.18.2",
    "pytest>=8.4.2",
    "pytest-asyncio>=0.24.0",
    "pytest-cov>=6.0.0",
    "ruff>=0.13.1",
]

[project.urls]
Homepage = "https://github.com/NimbleBrainInc/mcp-webfetch"
Repository = "https://github.com/NimbleBrainInc/mcp-webfetch.git"
Issues = "https://github.com/NimbleBrainInc/mcp-webfetch/issues"
